docs: update contributors
