+++ CONTRIBUTING.md
@@ -1,74 +1,91 @@
 # Contributing to Math W Code
 
-Thank you for considering contributing to Math W Code! We welcome contributions from everyone. By participating in this project, you agree to abide by the following guidelines.
+Hi There!👋 Thanks for your interest The team behind **Math W Code** appreciates your interest to help contribute in the project. We want to thank you for joining us, and encourage everyone to provide their input. So, let’s make this fun, creative and productive
+let’s build something great together. Just by agreeing to participate you are also agreeing to adhere to a couple of rules.
 
 ## Code of Conduct
 
-Please read and follow our [Code of Conduct](https://www.contributor-covenant.org/version/2/0/code_of_conduct/).
+First things first—please take a moment to review our [Code of Conduct](CODE_OF_CONDUCT). Let’s make sure this space is respectful and inclusive for all!
 
 ## How to Contribute
 
-### Reporting Content Errors
+### Found a Mistake? 💡
 
-If you find a mistake in any of our content or an improvement that can be made, please report it by opening an issue on our [GitHub repository](https://github.com/xirf/mathwithcode/issues). Include as much detail as possible to help us understand the issue.
+Oops, we have committed a mistake our content or perhaps you are a suggestion to make look even better? No worries! That is why you simply need to [open an issue](https://github.com/xirf/mathwithcode/issues). Include as much detail as possible so we can understand what needs fixing.
 
-### Suggesting Content
+### Want to Add Something New? 🚀
 
-If you have an idea for a new topic or a feature you would like to see, please open an issue on our [GitHub repository](https://github.com/xirf/mathwithcode/issues). We will review your suggestion and consider adding it to our content.
+Have an idea for a new topic or feature? Amazing! Feel free to open an issue on our [GitHub repository](https://github.com/xirf/mathwithcode/issues). If you prefer to jump straight into a pull request, that’s cool too! Just know that if there’s an issue already assigned to it, we might prioritize that first.
 
-### Submitting Changes
+<details>
+<summary>Want to Add a Fun Twist? 🐰</summary>
+
+Feel free to sneak in some fun (appropriate!) Easter eggs in the content you create! We love creativity and little surprises, but make sure they don’t distract from the main educational content.
+
+</details>
+
+### Ready to Submit Changes? 🛠️
+
+Here's a quick guide to submitting your contributions:
 
 1. Fork the repository.
-2. Create a new branch (`git checkout -b feature-branch`).
-3. Make your changes.
-4. Commit your changes (`git commit -m 'Add new feature'`).
-5. Push to the branch (`git push origin feature-branch`).
+2. Create a new branch (`git checkout -b your-feature-branch`).
+3. Work your magic and make your changes!
+4. Commit your work with a meaningful message (`git commit -m 'Add new feature'`).
+5. Push to your branch (`git push origin your-feature-branch`).
 6. Open a pull request on our [GitHub repository](https://github.com/xirf/mathwithcode/pulls).
 
-### Writing Standards
+And that’s it—you’re on your way to becoming a contributor!
 
-- Follow the existing code style.
-- Write clear, concise commit messages.
-- Include comments where necessary.
-- Ensure your content is easy to understand.
-- Use proper grammar and spelling as much as possible mistakes happen, but try to keep them to a minimum.
-- Use the appropriate language for the topic you are writing about, don't mix languages in the same document.
-- Everything in docs folder except `.vitepress` should be written in vuepress markdown  and should be in Indonesian language with English file/folder name.
-- File and folder names should be in English and use kebab-case.
-- Mixing languages in the same document is allowed, but it should be used sparingly and only when necessary.
-- Be creative! you can leave credit for yourself in the document you created in the bottom of the document. 
+### Keep It Neat! 📝
 
+When writing or coding, try to stick to these guidelines:
 
-### Crediting Yourself
-You can leave credit for yourself in the document you created in the bottom of the document. You can use the following template:
+- Follow the existing style of code (no need to reinvent the wheel 😉).
+- Write clear, concise commit messages that explain your changes.
+- Comment when necessary to explain complex parts.
+- Aim for simplicity and clarity—your work should be easy for others to follow.
+- Please use proper grammar and spelling. Everyone makes mistakes, but let’s keep them to a minimum.
+- Name your files and folders in English and using `kebab-case` (e.g., `linear-algebra`).
+- Be as descriptive as possible with your filenames and folder names (keep it short 👀).
+- Mixing languages in the same document is fine, but do it sparingly.
+- Get creative! You can leave a credit for yourself at the top of the document you created. ✨
+
+### Programming Language Preferences
+
+Ah, we like the simple and functional approach so much! Here are a few preferences:
+
+- Use `JavaScript` instead of `TypeScript`. No need for extra types.
+- Prefer `Kotlin` over `Java`. It’s cleaner and less boilerplate.
+- If possible skip the main function in `C++`, `Python` and more. We want to keep it simple.
+- Avoid adding lesser-known or overly niche programming languages unless necessary.
+
+### Crediting Yourself 🙌
+
+Want to give yourself some credit? Go for it! You can add a little shout-out in the front matter (top of document) of the document using this format:
 
 ```markdown
 ---
-# Your Name
-# Your Social Media
-# Your Website
+contributors:
+  - name: "Your Name"
+    username: "Your GitHub Username"
 ---
 ```
 
-You are free to leave any of the fields blank if you don't want to include them. but remember don't include any sensitive information or something that promote one of the following:
-- Hate speech
-- Discrimination
-- Harassment
-- Violence
-- Illegal activities (Indonesian law)
+Feel free to update the contributors' widget in `docs/.vitepress/theme/components/ContributorsAside.vue` to add more details about yourself—but make sure you don’t break the existing format! If it does break, it’s up to you to fix it. 😇
 
-Your credit may be removed or censor if it violates the guidelines.
+### The Review Process
 
-### Review Process
-
-All pull requests will be reviewed by the maintainers. We may ask for changes or improvements before merging your changes. We will do our best to provide feedback in a timely manner.
+Once you submit a pull request, one of our maintainers will review it. We may suggest a few tweaks, but don’t worry—we’ll provide feedback as quickly as possible and work with you to get your contributions merged!
 
 ### License
 
-By contributing to Math W Code, you agree that your contributions will be licensed under the MIT License.
+By contributing to Math W Code, you agree that your work will be licensed under the MIT License. This allows others to use, modify, and distribute your work freely. You can read the full license [here](LICENSE).
 
-## Contact
+## Questions? Need Help?
 
-If you have any questions, feel free to reach out to us at [contact@mathwcode.jamcoder.id](mailto:contact@mathwcode.jamcoder.id).
+Feel free to reach out to us at [itsandka@gmail.com](mailto:itsandka@gmail.com) if you have any questions or need assistance. We're happy to help!
 
-Thank you for your contributions!
+---
+
+Thank you for being part of our community! Let's build something great together. 🚀
