+++ README.md
@@ -1,51 +1,45 @@
-# Math W Code
+<div align="center">
+    <img src="docs/public/icon.png" alt="Math W Code Logo" width="100">
+    <h1>Math W Code</h1>
+    <p>
+        Learn math and it's implementation in computer programming. Learn everything from arithmetic to calculus with code written by many contributors from around the world.
+    </p>
+</div>
 
-Welcome to Math W Code! This is a website dedicated to learning math with code. Here, you can learn and contribute to the development of mathematical concepts through code!
+> [!WARNING] This project is still under development. 
+> This project is still in the early stages of development. We are working hard to make it better for everyone. If you have any suggestions or want to contribute, feel free to reach out to us.
+
+
 
 ## Table of Contents
 
-- [Math W Code](#math-w-code)
-  - [Table of Contents](#table-of-contents)
-  - [Introduction](#introduction)
-  - [Features](#features)
-  - [Installation](#installation)
-  - [Usage](#usage)
-  - [Contributing](#contributing)
-  - [License](#license)
-  - [Contact](#contact)
-  - [Acknowledgements](#acknowledgements)
+- [Table of Contents](#table-of-contents)
+- [Introduction](#introduction)
+- [Installation](#installation)
+- [Usage](#usage)
+- [Contributing](#contributing)
+- [License](#license)
+- [Contact](#contact)
+- [Thanks to Our Contributors](#thanks-to-our-contributors)
 
 ## Introduction
 
-Math W Code aims to make learning math fun and engaging by providing interactive coding exercises that help users understand mathematical concepts through practical application.
-
-## Features
-
-- Interactive coding exercises
-- Step-by-step tutorials
-- Comprehensive math topics
-- User-friendly interface
+Math W Code is a community-driven project that aims to help people learn math and its implementation in computer programming. We want everyone to learn math in a different way, by using code to help visualize and understand the concepts better. 
 
 ## Installation
 
 To set up the project locally, follow these steps:
 
-1. Clone the repository:
-    ```sh
-    git clone https://github.com/xirf/mathwithcode.git
-    ```
-2. Navigate to the project directory:
-    ```sh
-    cd mathwithcode
-    ```
-3. Install the dependencies:
-    ```sh
-    npm install
-    ```
+| Step | Command |
+|------|---------|
+| Clone the repository | ```sh git clone https://github.com/xirf/mathwithcode.git ``` |
+| Navigate to the project directory | ```sh cd mathwithcode ``` |
+| Install the dependencies | ```sh npm install ``` |
 
 ## Usage
 
 To start the development server, run:
+
 ```sh
 npm run dev
 ```
@@ -53,13 +47,17 @@ npm run dev
 Open your browser and go to `http://localhost:5173` to view the website locally.
 
 ## Contributing
+
 We welcome contributions from everyone. Please read our [Contribution Guidelines](CONTRIBUTING.md) and [Code of Conduct](CODE_OF_CONDUCT.md) before making a pull request just to make sure everyone is on the same page.
 
 ## License
+
 This project is licensed under the MIT License. See the LICENSE file for more details.
 
 ## Contact
+
 If you have any questions or suggestions, feel free to reach out to us at [contact@mathwcode.jamcoder.id](mailto:contact@mathwcode.jamcoder.id)
 
-## Acknowledgements
-This project was inspired by my ([@xirf](https://github.com/xirf)) passion for math and coding. He isn't the best at either, but he tries his best to make learning fun and engaging for everyone. Some content may not be accurate, so please feel free to contribute to the project and help make it better for everyone.
+## Thanks to Our Contributors
+
+A big shoutout to all our contributors who have helped make this project better. You can find the list of contributors [here](CONTRIBUTORS.md).
