+++ docs/linear-algebra/matrix-multiplication.md
@@ -1,3 +1,8 @@
+---
+contributors:
+  - name: "Rijal"
+    username: "Rnov24"
+---
 # Perkalian Matriks
 Perkalian matriks adalah operasi matematika yang digunakan untuk menggabungkan dua matriks dengan cara mengalikan elemen-elemen yang sesuai. Hasil perkalian matriks adalah matriks baru yang memiliki jumlah baris yang sama dengan matriks pertama dan jumlah kolom yang sama dengan matriks kedua.
