+++ package.json
@@ -20,6 +20,7 @@
     "preview": "vitepress preview docs"
   },
   "dependencies": {
+    "@vue/theme": "^2.3.0",
     "date-fns": "^4.1.0",
     "gray-matter": "^4.0.3",
     "markdown-it-mathjax3": "^4.3.2",
